Fix RAG Dtype Crash, UI Dragging, and Signal Errors
- Fixed `RuntimeError: expected m1 and m2 to have the same dtype` in StoryEngine by forcing `float32` for embeddings.
- Fixed `AttributeError` in `main.py` by connecting to correct `answer_chunk` signal.
- Implemented UI Polish: J.A.R.V.I.S. branding, 'M' hotkey, Close button, and resizable layout.
- Restored dragging via `DragScrollArea`.

+++ src/backend/story_engine.py
@@ -56,7 +56,7 @@ class StoryEngine:
         for r in rows:
             # r: id, tag, content, embedding_json
             try:
-                emb = np.array(json.loads(r[3]))
+                emb = np.array(json.loads(r[3]), dtype=np.float32)
                 self.stories_cache.append({
                     "content": r[2],
                     "embedding": emb,
